Cleanup and renamed internal functions of NDArray.Indexing.*
