+++ src/NumSharp.Core/Indexing/np.nonzero.cs
@@ -12,19 +12,15 @@ namespace NumSharp
 {
     public static partial class np
     {
-
         /// <summary>
-        /// Return the indices of the elements that are non-zero.
-        /// 
-        /// Returns a tuple of arrays, one for each dimension of a, containing the indices of the non-zero elements in that dimension.The values in a are always tested and returned in row-major, C-style order.
-        ///
-        /// To group the indices by element, rather than dimension, use argwhere, which returns a row for each non-zero element.
+        ///     Return the indices of the elements that are non-zero.
+        ///     Returns a tuple of arrays, one for each dimension of a, containing the indices of the non-zero elements in that dimension.The values in a are always tested and returned in row-major, C-style order.
+        ///     To group the indices by element, rather than dimension, use argwhere, which returns a row for each non-zero element.
         /// </summary>
-        /// <param name="a"></param>
-        /// <returns></returns>
+        /// <param name="a">Input array.</param>
+        /// <returns>Indices of elements that are non-zero.</returns>
+        /// <remarks>https://docs.scipy.org/doc/numpy/reference/generated/numpy.nonzero.html</remarks>
         public static NDArray<int>[] nonzero(in NDArray a)
             => a.TensorEngine.NonZero(a);
-
-
     }
 }
