+++ src/NumSharp.Core/Selection/NDArray.Indexing.Masking.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Diagnostics;
+using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Runtime.CompilerServices;
 using NumSharp.Backends;
@@ -16,9 +17,10 @@ namespace NumSharp
         /// <remarks>https://docs.scipy.org/doc/numpy-1.17.0/user/basics.indexing.html</remarks>
         /// <exception cref="IndexOutOfRangeException">When one of the indices exceeds limits.</exception>
         /// <exception cref="ArgumentException">indices must be of Int type (byte, u/short, u/int, u/long).</exception>
+        [SuppressMessage("ReSharper", "CoVariantArrayConversion")]
         public NDArray this[NDArray<bool> mask]
         {
-            get => retrieve_indices(this, np.nonzero(mask), null, true);
+            get => FetchIndices(this, np.nonzero(mask), null, true);
             set
             {
                 throw new NotImplementedException("Setter is not implemented yet");
